Publish data to S3 bucket (#6)
_Don't worry, the line count is misleading, this PR only adds about 100
lines of code_

This PR attempts to create a simple system that'll grow with the
project. As of this PR it...

1. enables an infrastructure-as-code system for AWS
2. creates an S3 bucket for our Canyoneer
3. uploads route beta to the S3 bucket
4. makes the S3 bucket available publicly

In the future we can use this same system to easily provision databases,
enable turnkey user authentication, create API endpoints, and more.

Please feel free to ask any questions you have and I'll try my best to
comment the code to make the unclear parts more clear!

Everyone on the project should have an AWS account invite waiting in
their email inbox.