Group all output products by schema version (#47)
fix #37 
fix #34 
fix #55 

This PR groups output products by their schema version (i.e.
`v2/index.json`). This will allow us to make breaking changes to the
schema in the future without breaking old clients.

It also makes some small changes to our data to reduce the
`v1/index.json` size. As of this PR it is 2.6 MB. For reference, the
`v1/index.json` is 24MB.

It also also adds some minor improvements to the README's description of
how to use the API