Add better integration test & test coverage metrics (#56)
This PR converts our one integration test for the `scrape` module into a
broader integration test for the CLI itself. It runs the CLI, ensures no
crashes, and compares the output products to a snapshot.

It also measures test coverage and ensures that test coverage does not
significantly decrease.

It also also renames the `--skipAWS` flag to `--local`... No particular
reason, I just think that's a better name.

Fix #29